Refactor: Apply dependency injection
dependency injection about search save service
